[Xamarin.Android.Build.Tasks] fix Debug mode and $(PublishTrimmed)
There is a problem if you have the combination:

* `Configuration=Debug`

* `PublishTrimmed=true`

We emit an `XA0119` warning with this combination, as there is not a
good reason to do it.

But unfortunately, the build will be completely broken as all the .NET
assemblies don't make it to the `.apk`! I could reproduce this in a
test.

The fix is that the logic in the `<ProcessAssemblies/>` MSBuild task
needs to match the logic in the `_PrepareAssemblies` MSBuild target:

    if (PublishTrimmed && !AndroidIncludeDebugSymbols) {
        //...
        ShrunkAssemblies = shrunkAssemblies.ToArray ();
    }

Should match:

    <ItemGroup Condition=" '$(PublishTrimmed)' == 'true' and '$(AndroidIncludeDebugSymbols)' != 'true' ">

With the preceding `<ItemGroup>` having the inverse of this `Condition`.

Going forward, we could probably refactor some of this logic to make
things simpler, but this is a reasonable fix for now.