[Xamarin.Android.Build.Tasks] make designer MSBuild targets a no-op
Context: https://developercommunity.visualstudio.com/t/XamarinAndroid-Designer---Replacement/10728132

As the Android designer is no longer supported in Visual Studio, we
can remove the MSBuild logic for the following targets:

* `GetExtraLibraryLocationsForDesigner`

* `SetupDependenciesForDesigner`

Since these are public MSBuild target names, we should probably leave
them in and "do nothing".

This also will simplify things if we ever refactor
`<GenerateJavaStubs/>` and friends as this removes a caller.

I also removed the `DesignerTests.cs` file that was testing these.