AndroidPackagingOptionsInclude is ignored (#9524)
Fixes #9520

PR #8459 added support for the `AndroidPackagingOptionsInclude`.
Unfortunately we forgot to add that property to the `BuildApk`
and `BuildBaseAppBundle` calls in the closed source part of
the product.

Since then we have move that code into this repo. So lets update all
the `BuildApk` and `BuildBaseAppBundle` calls to use the new property.