[Xamarin.Android.Build.Tasks] fix NRT warnings & other warnings (#9483)
Introduces the setting to `Xamarin.Android.Build.Tasks.csproj`:

    <WarningsAsErrors>Nullable</WarningsAsErrors>

Then I fixed the existing warnings. Down the road this will make it
easier to adopt `#enable nullable`, and eventually `Nullable=enable`
the entire project. Trying to do that now results in 1,000+ warnings...

Most of these are types that could easily be null, and they weren't
marked as nullable with `?`. Luckily, the code appears to be checking
for null in these places already.

For `*Attribute` types that need `IJniNameProviderAttribute.Name`,
I explicitly implemented a non-nullable version of this property in
`*Attribute.Partial.cs` files.