[build] produce multiple "flavors" of runtime packs
Right now our runtime packs are all identical for each .NET "runtime",
but we want to be able to produce different packs, such as:

    Microsoft.Android.Runtime.Mono.35.android-arm
    Microsoft.Android.Runtime.Mono.35.android-arm64
    Microsoft.Android.Runtime.Mono.35.android-x86
    Microsoft.Android.Runtime.Mono.35.android-x64
    Microsoft.Android.Runtime.NativeAOT.35.android-arm
    Microsoft.Android.Runtime.NativeAOT.35.android-arm64
    Microsoft.Android.Runtime.NativeAOT.35.android-x86
    Microsoft.Android.Runtime.NativeAOT.35.android-x64

Currently, the contents of all these packs are identical, but this is
an initial step to build multiple packs.

The plan is for the workload to only install the "Mono" packs, and the
"NativeAOT" packs can be restored via NuGet.