[Xamarin.Android.Build.Tasks] Remove AndroidUseAapt2 property (#9506)
Remove the `$(AndroidUseAapt2)` MSBuild property.  All of the code it
was protecting is now the default and users are no longer able to use
`aapt`; this property is redundant.

Update the `<ResolveAndroidTooling/>` task to look for `.cmd`, `.bat`,
and `.exe` like we do other tools.  While this is not needed at the
moment, it is useful for unit tests since we cannot create a `.cmd`
file and name it `.exe` and have it run.  It also allows us the
flexibility to write wrappers for the tool if needed.