[XABT] Break BuildApk into individual tasks for each content type. (#9612)
Refactor `BuildApk` into individual tasks for each content type.  This promotes easier understanding of which pieces are dependent on each other and on which MSBuild properties.

At this point, several of these tasks (like `CollectAssemblyFilesForArchive`) are still unconditionally creating and modifying files on every run. A future PR will separate these creations/modifications into separate _targets_ that can be run incrementally.

Theoretically, this PR is likely a slight performance regression due to having more tasks and a little bit of duplicated logic, however this should be more than offset by the ability to run tasks incrementally in the future. (The performance difference is small enough that it isn't really measurable because it is within the variability of our build time.)

| _BuildApkEmbed/_BuildApkFastDev target | `main` | This PR |
|--------|--------|--------|
| Debug (FastDev) | 2.674s | 2.655s |
| Debug (EmbedAssembliesIntoApk = true) | 64.384s | 64.357s |
| Release | 6.130s | 6.094s |