[Mono.Android] introduce `#if NATIVEAOT` for `monodroid_get_log_categories()`
In a NativeAOT context, we don't have `libmonodroid.so` *at all*.

This change introduces a new `#if NATIVEAOT` block in `Logger.cs` to
avoid calling `monodroid_get_log_categories()` and using `Default |
Assembly` by default.

I also reworked the build, so that the `$(AndroidRuntime)` property is
defined in `Configuration.props`. Other projects can eventually use it
as-needed.

In a future PR, we could consider p/invoke into
`__system_property_get()` and parse values. But we might want to
reconsider the *names* of some of the system properties as they have
`debug.mono.*` prefixes.

Hardcoding a default value is at least *better*, so it won't crash.