Prefer versions from android-platform-support
Context: 51151d727a0438dd049ad2678f9039d8f255d7b2

Commit 51151d72 laid out a conundrum: dotnet/android builds against
one set of versioned dependencies, but we want the "Xamarin Manifest"
to *actually* dictate the preferred package versions, as this makes
it easier to keep versions in sync with the VS installer/etc.

The problem is that commits prior to this point ignored the
`Xamarin.Installer.Common.props` described in 51151d72.

Additionally, many of the MSBuild properties are (1) "conditional" --
only setting a value if not overridden -- and (2) have the same
names as the properties within `Xamarin.Installer.Common.props`.

Square this circle by doing two things:

 1. Add `tools/workload-dependencies/WorkloadDependencies.proj`
    alongside `tools/workload-dependencies/Directory.Build.props`,
    which allows for an "isolated" environment, one in which
    *only* `Xamarin.Installer.Common.props` and
    `Xamarin.Android.Common.props` are imported.

    This allows us to cleanly ensure that
    `Xamarin.Installer.Common.props` controls things.

 2. Update `Microsoft.NET.Sdk.Android.proj` to use the
    `<MSBuild/>` task to invoke `WorkloadDependencies.proj`,
    providing all MSBuild properties required for things to work.

This should ensure that package versions are consistent.

Additionally, update `Program.cs` so that the `platform-tools`
package contains a `requestedRevision` property.
The `platform-tools` package is *unversioned*, so this is the only
way to get some semblance of versioning involved.