[Mono.Android] `Java.Lang.Object.GetObject<T>()` implementation
The current implementation:

    return Java.Interop.TypeManager.CreateInstance (handle, transfer, type);

Needs to become this, for NativeAOT support:

    JniObjectReference reference = JNIEnv.CreateJniObjectReference (handle, transfer);
    JniObjectReferenceOptions options = JNIEnv.ToJniObjectReferenceOptions (transfer);
    return (IJavaPeerable) JNIEnvInit.AndroidValueManager?.GetValue (ref reference, options, type);

This way we use the appropriate `AndroidValueManager` for the runtime
that is currently being used. If we just change the code in
`Java.Lang.Object` to do this for all runtimes, this can keep
everything a bit simpler.

The fallout of this change requires `type` to be decorated with:

    [DynamicallyAccessedMembers (Constructors)]

This trickled to create many other trimmer warnings, that are likely a
real issue. This is a good opportunity to fix them.

I also defined the `const` value in `Java.Lang.Object`:

    internal const DynamicallyAccessedMemberTypes Constructors = DynamicallyAccessedMemberTypes.PublicConstructors | DynamicallyAccessedMemberTypes.NonPublicConstructors;

Which allowed many types that extend `Java.Lang.Object` able to reuse
this `const` value and not define it again.

I also had to make a helper function to map `JniHandleOwnership` ->
`JniObjectReferenceOptions`, such as:

    internal static JniObjectReferenceOptions ToJniObjectReferenceOptions (JniHandleOwnership transfer)

And `CreateJniObjectReference()`:

    internal static JniObjectReference CreateJniObjectReference (IntPtr handle, JniHandleOwnership transfer)

Other changes:

* Bump to dotnet/java-interop/main@2c06b3c2a

Changes: https://github.com/dotnet/java-interop/compare/f800ea52...2c06b3c2a