[NativeAOT] improve build logic, part 1 (#9614)
Introduce `$(_AndroidNativeAot)` to be used throughout a build to
know if it is a NativeAOT build.  The logic for detecting a NativeAOT
build will likely change for Debug vs Release in the future.

Skip generation of `mono.MonoRuntimeProvider` for NativeAOT and
don't emit the value in `AndroidManifest.xml`.

Update `Microsoft.Android.Sdk.NativeAOT.targets` to use
`@(LinkerArg)` and `%(ResolvedFileToPublish.ArchiveFileName)` to
ensure that the NativeAOT-generated native library has a `lib` prefix,
as `System.loadLibrary("@NAME@")` only looks for `lib@NAME@.so`.

Update `BuildTest2.NativeAOT()` to ensure that the native library
`libHello.so` is created, and that `classes.dex` does NOT contain
`mono.MonoRuntimeProvider`.