Remove call to CollectAssemblyFilesForArchive for Fast Deployment.
Commit 6b91b042 split up the BuildApk task into their own Tasks to
improve build performance. In that change the call to
`CollectAssemblyFilesForArchive` was a no-op if `EmbedAssembliesIntoApk`
was set to `False`. However it was still being called in the `_BuildApkFastDev`
target, even though it had a condition which would skip the target
entirely if `EmbedAssembliesIntoApk` was `False`.

So lets remove the call to `CollectAssemblyFilesForArchive` from `_BuildApkFastDev`
completely. We can also remove the `EmbedAssemblies` property from the
`CollectAssemblyFilesForArchive` task as it is no longer needed.