Moar fixes!
Try dotnet/java-interop#1287, which contains a Gradle bump, which
is probably needed because previously Linux & Windows failed to build
because Gradle 8.1 and JDK-21 don't mix.

Update `$(LatestSupportedJavaVersion)` to 21.0.99 so that our unit
tests will actually *try* to do something under JDK-21, instead of
insta-failing with an XA0030 error:

> error XA0030: Building with JDK version `21.0.5` is not supported.
> Please install JDK version `17.0`. See https://aka.ms/xamarin/jdk9-errors

Doh.

Update to Gradle 8.12, for harmony with dotnet/java-interop#1287.