[XABT] Remove <CollectAssemblyFilesForArchive/> from FastDeployment (#9650)
Context: 6b91b0423dab56907b1d1dc41677532f7e44f8e3

Commit 6b91b042 split up the `<BuildApk/>` task into multiple
`<Collect*/>` tasks to improve build performance.  In that change the
call to `<CollectAssemblyFilesForArchive/>` was a no-op if
`CollectAssemblyFilesForArchive.EmbedAssembliesIntoApk` is `False`.
However it was still being called in the `_BuildApkFastDev` target,
even though it had a condition which would skip the target entirely
when `$(EmbedAssembliesIntoApk)=False`.

Remove the call to `<CollectAssemblyFilesForArchive/>` from
`_BuildApkFastDev` completely.  We can also remove the
`CollectAssemblyFilesForArchive.EmbedAssemblies` property from the
task as it is no longer needed.