[Xamarin.Android.Build.Tasks] Add %(NuGetPackage*) to TaskItems (#9559)
Fixes: https://github.com/dotnet/android/issues/9544

Context: 64bb147f2ead5bd403a823d27b33f590197477fa

[Google announced][0] that future versions of Android would require
that native libraries use 16 KB page sizes on arm64.  At present, the
timeline for *when* 16 KB page sizes will be required is unknown,
though we assume it will be with Android 16 or later.

In order to get ahead of this, .NET 9 added an XA0141 warning in
64bb147f.

The problem is, this warning is not entirely actionable:

	dotnet new android
	dotnet add package Xamarin.GooglePlayServices.Vision.Face.Contour.Internal --version 116.1.0.19
	dotnet build -c Release

results in:

	warning XA0141: NuGet package '<unknown>' version '<unknown>' contains a shared library 'libface_detector_v2_jni.so'
	  which is not correctly aligned. See https://developer.android.com/guide/practices/page-sizes for more details
	warning XA0141: NuGet package '<unknown>' version '<unknown>' contains a shared library 'libface_detector_v2_jni.so'
	  which is not correctly aligned. See https://developer.android.com/guide/practices/page-sizes for more details

What are customers supposed to *do* with this?

The original assumption was that the NuGet package and Version would
be available as metadata items on the MSBuild Items.  Unfortunately
that was not the case, so the data the user gets is empty.

Add the required metadata to all NuGet Package references resolved by
the project.  This will allow us to propagate that data throughout
the build Items as they are transformed.  This means we can provide a
decent error message to the user:

	warning XA0141: Android 16 will require 16 KB page sizes, shared library 'libface_detector_v2_jni.so' does not have a 16 KB page size.
	  Please inform the authors of the NuGet package 'Xamarin.GooglePlayServices.Vision.Face.Contour.Internal' version '116.1.0.19'
	  which contains 'lib/net8.0-android34.0/play-services-vision-face-contour-internal.aar'.
	  See https://developer.android.com/guide/practices/page-sizes for more details.

[0]: https://android-developers.googleblog.com/2024/08/adding-16-kb-page-size-to-android.html