[tests] use the 'TestName' property (#9664)
Context: https://github.com/dotnet/msbuild/issues/11237

In .NET 10, we noticed there is a bug if a project contains `(` or `)` in the path.

We had various tests that were not using `TestName` to sanitize the path.

We have existing tests that do test `(` and `)` characters, which we have left in place.