[ci] Run Mac MSBuild tests on `macos-14-arm64` (#9665)
Azure Pipelines now provides the `macos-14-arm64` image.  We can use this image to run our MSBuild tests on.  This is a benefit because it's nearly twice as fast and nowadays our Mac users are probably more likely to be running arm64.

Unfortunately it [does not support nested virtualization](https://github.com/actions/runner-images/issues/9460), so we can't run our emulator tests on this image.  Supposedly Apple added nested virtualization in MacOS 15, but Azure does not have `macos-15-arm64` images yet.  Hopefully we will be able to use those for emulator tests in the future.

Additionally:
- Update `apkdiff` to a [newer version with Apple Silicon support](https://github.com/radekdoulik/apkdiff/pull/8).
- Tweak agent count since we now need fewer Mac agents, but could use a couple more Windows agents.
- Bugfix to make sure we're using the right JDK path for ARM64.