[XABT] Create separate assembly preparation tasks. (#9637)
Break down the `CollectAssemblyFilesForArchive` task further into individual tasks.

These tasks each handle a separate concern about preparing the assemblies for the apk:

- `CompressAssemblies` - Compresses assemblies using LZ4 compression (independent of ZIP compression).
- `CreateAssemblyStore` - Bundles assemblies into a single `.blob` file.
- `WrapAssembliesAsSharedLibraries` - Wraps the assemblies as shared libraries so they can go in the `lib` directory of the APK.

This should allow them to be moved to separate incremental targets in the future. For example, we do not need to compress all assemblies on every build if they haven't changed since the previous build.