[Xamarin.Android.Build.Tasks] Rework DTB to use .aar files directly (#9700)
Context: https://developer.android.com/studio/projects/android-library#aar-contents

One of the big issues we have with design time builds is the shear
amount of time they take.  This is principally because we have to
discover and extract *all* the `.aar`/`.zip` files which we need
before we can do a build.

This used to be because we embedded the data into the assemblies, but
we ship these resources as `.aar` file these days.
Let's take advantage of that.

Rather then extracting to disk and then reading the files from disk,
we instead read the `.aar` directly for a design time build.
We extract the required data into memory which will allow us to
generate the required `R.txt` file.  Note that only a small amount of
data is needed from the `.aar`, its mostly entry filenames and the
contents of any `.xml`/`.axml` files which we find in the `res` folder.
This will cut down the amount of data we need.

We only use this for design time builds.

The typical build is left as-is, and if the on disk cache is available
the design time build will use that instead of the `.aar`s.

One of the things that was missing from our design time build tests
was the `SkipCompilerExecution` MSBuild parameter.  This is one of
the important parameters which VS uses, as it stops the `<Csc/>` Task
from creating an assembly on disk.  This was the main speed issue we
had in our unit tests, and adding this flag cut a second off of our
design time build times. 


|               |  Average Time |
| ------------- | ------------: |
|      ef9912af |   2026.2504ms |
|   This Commit |    592.1775ms |