[ci] Set `$(TreatWarningsAsErrors)` = `true` for CI builds. (#9689)
Modern best practices for software development recommend treating
warnings as errors.  Although the majority of warnings are benign,
warnings for any potential actual issues will be lost in the sea of
benign warnings and will not be caught.

However, having to fix warnings during the development process can be
annoying as code is often in a temporary state.  As a balance, we
only error on warnings when building on CI.  This way local
development isn't hindered, but CI will protect us from committing
new warnings.

Developers can opt-in to the errors locally by setting
`$(_AndroidTreatWarningsAsErrors)` to `true`.

There are several projects that have warnings today.  For now, we
will simply grandfather those projects in, allowing us to move
forward with protecting the rest of the projects from new warnings.
With time, we should endeavor to fix these projects' warnings and
remove them from this list.  (Our build currently has ~170 warnings
spread across these projects.)